refactor copy-pasta into its own function
