+++ application/react/_hooks/useAxios.ts
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import axios from '../axios';
+import axios, { setBearerTokenToAxiosRequestsUsingInterceptors } from '../axios';
 import { useAuth } from './useAuth';
 import { useRefreshToken } from './useRefreshToken';
 import { AxiosInstance } from 'axios';
@@ -10,34 +10,7 @@ export const useAxios = (): AxiosInstance => {
   const { auth }: { auth: UserAuth} = useAuth();
 
   useEffect(() => {
-    const requestIntercept = axios.interceptors.request.use(
-      config => {
-        if (!config.headers['Authorization']) {
-          config.headers['Authorization'] = `Bearer ${auth?.accessToken}`;
-        }
-        return config;
-      }, (error) => Promise.reject(error)
-    );
-
-    const responseIntercept = axios.interceptors.response.use(
-      response => response,
-      async (error) => {
-        const prevRequest = error?.config;
-
-        if (error?.response?.status === 403 && !prevRequest?.sent) {
-          prevRequest.sent = true;
-          const newAccessToken = await refreshAccessToken();
-          prevRequest.headers['Authorization'] = `Bearer ${newAccessToken}`;
-          return axios(prevRequest);
-        }
-        return Promise.reject(error);
-      }
-    );
-
-    return () => {
-      axios.interceptors.request.eject(requestIntercept);
-      axios.interceptors.response.eject(responseIntercept);
-    }
+    setBearerTokenToAxiosRequestsUsingInterceptors(axios, auth, refreshAccessToken);
   }, [auth, refreshAccessToken])
